src/requirements.txt
flask
transformers
torch
pytorch-lightning
fairscale
numba
einops
mashumaro
pandas
bio
rdkit
ihm
modelcif
scipy
google-cloud-storage
gunicorn